+++ PingUI/Views/EditTargetView.axaml.cs
@@ -1,3 +1,4 @@
+using Avalonia.Controls;
 using Avalonia.ReactiveUI;
 using PingUI.ViewModels;
 
@@ -8,5 +9,12 @@ public partial class EditTargetView : ReactiveUserControl<EditTargetViewModel>
 	public EditTargetView()
 	{
 		InitializeComponent();
+		AddressText.AttachedToVisualTree += (sender, e) =>
+		{
+			if (sender is TextBox textBox)
+			{
+				textBox.Focus();
+			}
+		};
 	}
 }